// stocks-monorepo/shared/types/quote.ts
export interface Quote {
  symbol: string;
  name?: string;
  price: number;
  change: number;
  changePercentage: number;
  volume: number;
  dayLow: number;
  dayHigh: number;
  yearHigh: number;
  yearLow: number;
  marketCap: number;
  priceAvg50: number;
  priceAvg200: number;
  exchange: string;
  open: number;
  previousClose: number;
  timestamp: number;
}